Enhanced development features.
Since the sources use the AMD/Intel/GCC names for the intrinsics,
and since that may very from time-to-time and site-to-site;
configure recognizes alternate header names in SSE2_SIMD_HEADER
and in SSSE3_SIMD_HEADER environment variables.
(Only effects argon2 currently, blake2 continues happly unaware.)

The configure/makefile generation supports writing custome include
headers to re-define the A.I.C. operation names used in the source
to that used to program the site's vector processor.
(Not yet put to use, but the support is there.)

Here I give away my age . . .
The configure script accepts an --enable-listing feature flag.
A complete, mixed source and assembly, *.lst file for each *.o
Handy when porting to a non-GCC and/or non-AMD/Intel vector processor.
I imagine it will also be handy for those who like to audit their
security related software.

Out-of-tree builds are fully supported, with generated files put
in the build directory rather than clutter up the source directory.

Cross-compiling not (yet) tested, but the intent is to support it.